fix(dart): dart:stable in Dockerfile + type-safe Duration backoff
dart:3.3 image lagged behind the stable SDK used in CI unit-test step,
causing AOT compile failures when package constraints resolved to newer
versions. dart:stable aligns Docker build with setup-dart@v1 SDK version.

Also tightened retryDelay backoff to use attempt.toDouble() for explicit
num typing, removing int-vs-num ambiguity across SDK patch versions.

+++ dart_ingestion/lib/adapters/statsbomb_adapter.dart
@@ -55,7 +55,7 @@ class StatsBombAdapter implements DataAdapter {
 
       // 5xx: server error — retry with linear backoff
       if (attempt < maxRetries) {
-        await Future.delayed(retryDelay * attempt);
+        await Future.delayed(retryDelay * attempt.toDouble());
       } else {
         throw Exception(
             'StatsBomb fetch failed for match $matchId: HTTP ${response.statusCode} '
